Add device check before preparing USB drive
Refactor USB preparation to check device before executing commands.

+++ create_bootable_media.sh
@@ -3,10 +3,12 @@
 lsblk
 
 
-#prepare the usb drive
-wipefs --all --force /dev/sda
-mkfs.vfat -F32 /dev/sda
-
+if (usb device == /dev/sda)
+{
+  #prepare the usb drive
+  wipefs --all --force /dev/sda
+  mkfs.vfat -F32 /dev/sda
+}
 
 
 
@@ -23,6 +25,7 @@ cd void-mklive
 
 
 #run mklive to build an iso and add custom modules
+#crypt, disk, wayland, nano, pipewire H/b-top etc
 ./mklive.sh -o void-custom3.iso -p "git nano NetworkManager wpa_supplicant tree mc fastfetch htop btop xorg-minimal vsv firefox qt6-wayland xdg-desktop-portal-gtk xdg-user-dirs-gtk wireplumber seatd mutter wayland-utils xf86-video-intel kde-baseapps kde-plasma metalog elogind breeze alsa-plugins-pulseaudio alsa-tools kpipewire pipewire wireplumber cryptsetup grub-x86_64-efi mesa-opencl mesa-vulkan-intel lvm2 btrfs-progs dosfstools kbd xtools"
